Started writing tests, 1 working, 1 WIP.
